requirements.txt
aiofiles==0.5.0
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
contextvars==2.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.7.22
httptools==0.1.1
httpx==0.11.1
hyperframe==5.2.0
idna==2.10
idna-ssl==1.1.0
immutables==0.14
multidict==4.7.6
psycopg2-binary==2.8.5
PyJWT==1.7.1
rfc3986==1.4.0
sanic==20.6.3
sniffio==1.1.0
typing-extensions==3.7.4.2
ujson==3.0.0
urllib3==1.25.10
uvloop==0.14.0
websockets==8.1
yarl==1.5.0
